+++ @types/settings.ts
@@ -1,2 +1,7 @@
-export type serverSettingsType = {};
+export type serverSettingsType = {
+  customSearch: {
+    name: string;
+    url: string; //検索キーワードが%sに入る
+  }[];
+};
 export type localSettings = {};
